refactor: enhance AICLIService by integrating AICLIMessageHandler
- Replaced manual session message buffer initialization with a centralized method from AICLIMessageHandler for consistency and maintainability.
- Streamlined response processing by delegating logic to AICLIMessageHandler, improving clarity and reducing code duplication.
- Removed outdated handling methods, ensuring message processing is now managed through the new handler.
- Updated tests to reflect changes in message handling and buffer management.

These modifications improve the overall structure and readability of the AICLIService, aligning with best practices for code organization.
